# hand-tuned kernels for small batch sizes