--- a/src/neural/cuda/layers.cc
+++ b/src/neural/cuda/layers.cc
@@ -142,13 +142,13 @@ void ConvLayer<half>::LoadWeights(float* pfilter, float* pBias, void* scratch) {
   // and then do the type / layout conversion using a kernel.
   assert(scratch);
   ReportCUDAErrors(
-      cudaMemcpyAsync(scratch, pfilter, weight_size, cudaMemcpyHostToDevice));
+      cudaMemcpy(scratch, pfilter, weight_size, cudaMemcpyHostToDevice));
   fp32NCHWtofp16NHWC((half*)weights, (float*)scratch, C, c_input_, C, c_input_,
                      filter_size_, filter_size_);
 
   if (pBias) {
     ReportCUDAErrors(
-        cudaMemcpyAsync(scratch, pBias, blas_size, cudaMemcpyHostToDevice));
+        cudaMemcpy(scratch, pBias, blas_size, cudaMemcpyHostToDevice));
 
     copyTypeConverted((half*)biases, (float*)scratch, C);
   }
@@ -161,11 +161,11 @@ void ConvLayer<float>::LoadWeights(float* pfilter, float* pBias,
       sizeof(float) * c_input_ * C * filter_size_ * filter_size_;
   size_t blas_size = sizeof(float) * C;
   ReportCUDAErrors(
-      cudaMemcpyAsync(weights, pfilter, weight_size, cudaMemcpyHostToDevice));
+      cudaMemcpy(weights, pfilter, weight_size, cudaMemcpyHostToDevice));
 
   if (pBias) {
     ReportCUDAErrors(
-        cudaMemcpyAsync(biases, pBias, blas_size, cudaMemcpyHostToDevice));
+        cudaMemcpy(biases, pBias, blas_size, cudaMemcpyHostToDevice));
   } else {
     ReportCUDAErrors(cudaMemset(biases, blas_size, 0));
   }
@@ -240,9 +240,9 @@ template <typename DataType>
 void BNLayer<DataType>::LoadWeights(float* cpuMeans, float* cpuVar) {
   size_t weight_size = sizeof(float) * C;
   ReportCUDAErrors(
-      cudaMemcpyAsync(means_, cpuMeans, weight_size, cudaMemcpyHostToDevice));
+      cudaMemcpy(means_, cpuMeans, weight_size, cudaMemcpyHostToDevice));
   ReportCUDAErrors(
-      cudaMemcpyAsync(variances_, cpuVar, weight_size, cudaMemcpyHostToDevice));
+      cudaMemcpy(variances_, cpuVar, weight_size, cudaMemcpyHostToDevice));
 }
 
 template <>
@@ -300,24 +300,23 @@ void SELayer<float>::LoadWeights(float* w1, float* b1, float* w2, float* b2,
   size_t weight_size2 = 2 * weight_size1;
 
   // Weight for the first FC layer.
-  ReportCUDAErrors(
-      cudaMemcpyAsync(w1_, w1, weight_size1, cudaMemcpyHostToDevice));
+  ReportCUDAErrors(cudaMemcpy(w1_, w1, weight_size1, cudaMemcpyHostToDevice));
 
   // Weight for the second FC layer.
-  ReportCUDAErrors(
-      cudaMemcpyAsync(w2_, w2, weight_size2, cudaMemcpyHostToDevice));
+  ReportCUDAErrors(cudaMemcpy(w2_, w2, weight_size2, cudaMemcpyHostToDevice));
 
   // Bias for the first FC layer.
-  ReportCUDAErrors(cudaMemcpyAsync(b1_, b1, numFc1Out_ * sizeof(float),
+  ReportCUDAErrors(
+      cudaMemcpy(b1_, b1, numFc1Out_ * sizeof(float),
                                    cudaMemcpyHostToDevice));
 
   // Bias for the second FC layer.
   ReportCUDAErrors(
-      cudaMemcpyAsync(b2_, b2, 2 * C * sizeof(float), cudaMemcpyHostToDevice));
+      cudaMemcpy(b2_, b2, 2 * C * sizeof(float), cudaMemcpyHostToDevice));
 
   // Bias for previous layer (Convolution).
   if (prevLayerBias) {
-    ReportCUDAErrors(cudaMemcpyAsync(bPrev_, prevLayerBias, C * sizeof(float),
+    ReportCUDAErrors(cudaMemcpy(bPrev_, prevLayerBias, C * sizeof(float),
                                      cudaMemcpyHostToDevice));
   }
 }
@@ -342,38 +341,41 @@ void SELayer<half>::LoadWeights(float* w1, float* b1, float* w2, float* b2,
   // Weight for the first FC layer.
   if (kUseFusedSELayer) {
     cpuTranspose(temp.data(), w1, numFc1Out_, C);
-    ReportCUDAErrors(cudaMemcpyAsync(scratch, temp.data(), weight_size1,
+    ReportCUDAErrors(
+        cudaMemcpy(scratch, temp.data(), weight_size1,
                                      cudaMemcpyHostToDevice));
   } else {
     ReportCUDAErrors(
-        cudaMemcpyAsync(scratch, w1, weight_size1, cudaMemcpyHostToDevice));
+        cudaMemcpy(scratch, w1, weight_size1, cudaMemcpyHostToDevice));
   }
   copyTypeConverted((half*)w1_, (float*)scratch, num_weights1);
 
   // Weight for the second FC layer.
   if (kUseFusedSELayer) {
     cpuTranspose(temp.data(), w2, 2 * C, numFc1Out_);
-    ReportCUDAErrors(cudaMemcpyAsync(scratch, temp.data(), weight_size2,
+    ReportCUDAErrors(
+        cudaMemcpy(scratch, temp.data(), weight_size2,
                                      cudaMemcpyHostToDevice));
   } else {
     ReportCUDAErrors(
-        cudaMemcpyAsync(scratch, w2, weight_size2, cudaMemcpyHostToDevice));
+        cudaMemcpy(scratch, w2, weight_size2, cudaMemcpyHostToDevice));
   }
   copyTypeConverted((half*)w2_, (float*)scratch, num_weights2);
 
   // Bias for the first FC layer.
-  ReportCUDAErrors(cudaMemcpyAsync(scratch, b1, numFc1Out_ * sizeof(float),
+  ReportCUDAErrors(cudaMemcpy(scratch, b1, numFc1Out_ * sizeof(float),
                                    cudaMemcpyHostToDevice));
   copyTypeConverted((half*)b1_, (float*)scratch, numFc1Out_);
 
   // Bias for the second FC layer.
-  ReportCUDAErrors(cudaMemcpyAsync(scratch, b2, 2 * C * sizeof(float),
+  ReportCUDAErrors(
+      cudaMemcpy(scratch, b2, 2 * C * sizeof(float),
                                    cudaMemcpyHostToDevice));
   copyTypeConverted((half*)b2_, (float*)scratch, 2 * C);
 
   // Bias for previous layer (Convolution).
   if (prevLayerBias) {
-    ReportCUDAErrors(cudaMemcpyAsync(scratch, prevLayerBias, C * sizeof(float),
+    ReportCUDAErrors(cudaMemcpy(scratch, prevLayerBias, C * sizeof(float),
                                      cudaMemcpyHostToDevice));
     copyTypeConverted((half*)bPrev_, (float*)scratch, C);
   }
@@ -482,7 +484,7 @@ void FCLayer<half>::LoadWeights(float* cpuWeight, float* cpuBias,
   // also need to convert from fp32 to fp16
   assert(scratch);
   ReportCUDAErrors(
-      cudaMemcpyAsync(scratch, cpuWeight, weight_size, cudaMemcpyHostToDevice));
+      cudaMemcpy(scratch, cpuWeight, weight_size, cudaMemcpyHostToDevice));
 
   fp32NCHWtofp16NHWC((half*)weights_, (float*)scratch, num_biases,
                      input_->GetC(), num_biases, input_->GetC(), input_->GetH(),
@@ -490,7 +492,7 @@ void FCLayer<half>::LoadWeights(float* cpuWeight, float* cpuBias,
 
   if (cpuBias) {
     ReportCUDAErrors(
-        cudaMemcpyAsync(scratch, cpuBias, blas_size, cudaMemcpyHostToDevice));
+        cudaMemcpy(scratch, cpuBias, blas_size, cudaMemcpyHostToDevice));
     copyTypeConverted((half*)biases_, (float*)scratch, num_biases);
   }
 }
@@ -504,11 +506,12 @@ void FCLayer<float>::LoadWeights(float* cpuWeight, float* cpuBias,
   size_t num_biases = C * H * W;
   size_t blas_size = sizeof(float) * num_biases;
 
-  ReportCUDAErrors(cudaMemcpyAsync(weights_, cpuWeight, weight_size,
+  ReportCUDAErrors(
+      cudaMemcpy(weights_, cpuWeight, weight_size,
                                    cudaMemcpyHostToDevice));
   if (use_bias_) {
     ReportCUDAErrors(
-        cudaMemcpyAsync(biases_, cpuBias, blas_size, cudaMemcpyHostToDevice));
+        cudaMemcpy(biases_, cpuBias, blas_size, cudaMemcpyHostToDevice));
   }
 }
 
@@ -567,7 +570,7 @@ FCLayer<DataType>::~FCLayer() {
 template <typename DataType>
 PolicyMapLayer<DataType>::PolicyMapLayer(BaseLayer<DataType>* ip, int C, int H,
                                          int W, int usedSize)
-    : BaseLayer<DataType>(C, H, W, ip), usedSize(usedSize) {
+    : BaseLayer<DataType>(C, H, W, ip), used_size_(usedSize) {
   size_t weight_size = sizeof(short) * this->input_->GetC() * 64;
   ReportCUDAErrors(cudaMalloc(&weights_, weight_size));
 }
@@ -575,14 +578,61 @@ PolicyMapLayer<DataType>::PolicyMapLayer(BaseLayer<DataType>* ip, int C, int H,
 template <typename DataType>
 void PolicyMapLayer<DataType>::LoadWeights(const short* cpuWeight,
                                            void* /*scratch*/) {
-  size_t weight_size = sizeof(short) * usedSize;
+  size_t weight_size = sizeof(short) * used_size_;
 
   if (std::is_same<half, DataType>::value) {
     // convert CHW to HWC
-    int C = usedSize / 64;
+    int C = used_size_ / 64;
     int Cin = this->input_->GetC();
-    usedSize = Cin * 64;
-    short* convertedWeights = new short[usedSize];
+
+    // C is the no. of channels actually used (typically 73).
+    // Cin the the no. of channels in previous layer (padded up to 80).
+    // Weights of this layer is a mapping to select which output index of the
+    // policy vector (1858 elements) maps to every element of input
+    // tensor (assuming NCHW layout). Note that there are 73x64 valid inputs
+    // (80x64 taking padding), and only 1858 outputs so the mapping isn't
+    // one to one. Only few of the indices point to valid index in policy
+    // vector. Invalid entries are set to -1.
+
+    // In fp16 mode, the tensor layout is NHWC so the weights need to be adjusted
+    // to make them work as intended.
+
+    // This is how the original weights looks like (CHW layout):
+    /*
+               HW (64)
+       ----|-------------|
+           |             |
+           |             |
+    C (73) |             |
+           |             |    
+           |             |
+       ------------------|   Cin (80)
+           |  padding    |
+           |-------------|
+    */
+    // The padding is not part of the weights provided (used_size_ is 73 x 64).
+    //
+    // The weights converted to HWC looks like this
+    /*
+                 C (73)
+            |-------------|---|
+            |             | P |
+            |             | a |
+    HW (64) |             | d |
+            |             |   |  
+            |             |   |
+            |-----------------| 
+                     Cin (80)    
+    */
+    // In HWC, because the padding is now part of each row
+    // we need to increase the size of weights mappnig to account
+    // for it. 
+    // The pad elements point to -1 (invalid output index) and the 
+    // same kernel works for both HWC and CHW layouts after used_size_
+    // is updated to include padding (80x64).
+
+    used_size_ = Cin * 64;
+    std::vector<short> convertedWeights(used_size_);
 
     for (int hw = 0; hw < 64; hw++)
       for (int c = 0; c < Cin; c++) {
@@ -591,12 +641,12 @@ void PolicyMapLayer<DataType>::LoadWeights(const short* cpuWeight,
         else
           convertedWeights[hw * Cin + c] = -1;
       }
-    ReportCUDAErrors(cudaMemcpyAsync(weights_, convertedWeights,
-                                     usedSize * sizeof(short),
+    ReportCUDAErrors(cudaMemcpy(weights_, &convertedWeights[0],
+                                     used_size_ * sizeof(short),
                                      cudaMemcpyHostToDevice));
-    delete[] convertedWeights;
   } else {
-    ReportCUDAErrors(cudaMemcpyAsync(weights_, cpuWeight, weight_size,
+    ReportCUDAErrors(
+        cudaMemcpy(weights_, cpuWeight, weight_size,
                                      cudaMemcpyHostToDevice));
   }
 }
@@ -611,7 +661,7 @@ void PolicyMapLayer<DataType>::Eval(int N, DataType* output_tensor,
   int inputSize =
       this->input_->GetC() * this->input_->GetH() * this->input_->GetW();
   int outputSize = this->C * this->H * this->W;
-  PolicyMap(N, output_tensor, input_tensor, weights_, inputSize, usedSize,
+  PolicyMap(N, output_tensor, input_tensor, weights_, inputSize, used_size_,
             outputSize);
 }