… src/neural/cuda/layers.h