Added a couple of hand-tuned kernels for cudnn backend for very small batch sizes (1-8). Quite a bit faster than cudnn kernels (but still very slow compared to performance with bigger batches).

here is some raw NPS dumps on RTX 2060 (taken with --max-prefetch=0, otherwise the search will try to send bigger batches than the --minibatch-size setting):
`
baseline:
---------
Avg batch size: 1, NN eval time: 0.701861 seconds per 100 evals. NPS: 142.478
Avg batch size: 1.96, NN eval time: 0.716327 seconds per 196 evals. NPS: 273.618
Avg batch size: 2.88, NN eval time: 0.717454 seconds per 288 evals. NPS: 401.42
Avg batch size: 3.93, NN eval time: 0.711275 seconds per 393 evals. NPS: 552.529
Avg batch size: 4.97, NN eval time: 0.709814 seconds per 497 evals. NPS: 700.183
Avg batch size: 5.91, NN eval time: 0.723776 seconds per 591 evals. NPS: 816.551
Avg batch size: 6.91, NN eval time: 0.721896 seconds per 691 evals. NPS: 957.202
Avg batch size: 7.83, NN eval time: 0.71533 seconds per 783 evals. NPS: 1094.6
Avg batch size: 9.79, NN eval time: 0.718194 seconds per 979 evals. NPS: 1363.14

Avg batch size: 16, NN eval time: 0.731791 seconds per 1600 evals. NPS: 2186.42


optimized:
----------
Avg batch size: 1, NN eval time: 0.138984 seconds per 100 evals. NPS: 719.505
Avg batch size: 1.92, NN eval time: 0.196821 seconds per 192 evals. NPS: 975.505
Avg batch size: 2.87, NN eval time: 0.262773 seconds per 287 evals. NPS: 1092.2
Avg batch size: 3.87, NN eval time: 0.330469 seconds per 387 evals. NPS: 1171.06
Avg batch size: 4.92, NN eval time: 0.405732 seconds per 492 evals. NPS: 1212.62
Avg batch size: 5.98, NN eval time: 0.480324 seconds per 598 evals. NPS: 1244.99
Avg batch size: 6.88, NN eval time: 0.546159 seconds per 688 evals. NPS: 1259.71
Avg batch size: 7.83, NN eval time: 0.609935 seconds per 783 evals. NPS: 1283.74
Avg batch size: 9.67, NN eval time: 0.701159 seconds per 967 evals. NPS: 1379.14


Avg batch size: 15.93, NN eval time: 0.730636 seconds per 1593 evals. NPS: 2180.29`


Additional data comparing different GPUs for batch size of 1:
`
GTX 970:
baseline: Avg batch size: 1, NN eval time: 1.66256 seconds per 100 evals. NPS: 60.1483
optimized: Avg batch size: 1, NN eval time: 0.268603 seconds per 100 evals. NPS: 372.297

RTX Titan:
baseline: Avg batch size: 1, NN eval time: 0.373341 seconds per 100 evals. NPS: 267.852
optimized: Avg batch size: 1, NN eval time: 0.117504 seconds per 100 evals. NPS: 851.034`


For small batches, OpenCL backend was much faster before. With these new kernels, it's slightly faster than OpenCL backend too:
minibatch-size of 1, on RTX 2060:
`OpenCL : Benchmark final time 8.71946s calculating 558.062 nodes per second.
cuda-opt: Benchmark final time 7.97152s calculating 966.064 nodes per second.`