@oscardssmith 2019-02-21
Can you do an elo test of batchsize 4 vs the standard with this optimization? I'd be interested to see if the tree is enough better shaped to cancel out the worse nps.