@ankan-ban 2019-02-21
Result from a short test with --minibatch-size=4 --max-prefetch=0 (TC 10+0.5 seconds on RTX 2060):
`Lc0-opt - Lc0-baseline : 33.5/45 23-1-21 (==1=====11=1=1==1=11=1=11111=1=1=111==01=1=11)  74%  +182`
(baseline is default cudnn backend)