@ankan-ban 2019-02-21
Very early results of running with default vs batch-size of 4:

`Lc0-opt - Lc0-baseline : 3.0/13 0-7-6 (0====00=0=000)  23%  -210`

Lc0-opt is --minibatch-size=4 --max-prefetch=0 (running new hand-tuned kernels)
Lc0-baseline is default batch size setting (--minibatch-size=256 --max-prefetch=32)

As expected default batch-size (256) is much stronger due to ~5x higher nps.
(I will update the results once I have more games)