@ankan-ban 2020-04-28
After the recent changes to cudnn backend (custom_winograd convolution implementation), the cudnn backend is faster than before - on par with OpenCL:

```
with v0.25 build, LS14 network, --minibatch-size=1, --max-prefetch=0 

OpenCL:
Benchmark time 10000ms, 5095 nodes, 515 nps, move e2e4

cudnn (default: custom_winograd=true):
Benchmark time 10000ms, 5241 nodes, 524 nps, move e2e4

cudnn (--backend-opts=custom_winograd=false):
Benchmark time 10000ms, 3004 nodes, 300 nps, move e2e4

dx12 (fp16=false)
Benchmark time 10000ms, 4498 nodes, 452 nps, move e2e4

cudnn (with this PR)
Benchmark time 10000ms, 8126 nodes, 813 nps, move e2e4
```

Which means this PR could still provide a good boost in performance for people doing 1node tests.