@ankan-ban 2020-04-28
Adding not for merge label. 
consenconsen at discord ran into functional and performance issues with this PR with GTX 1650, so it's possible that it doesn't work well on all GPUs, and with all network sizes. More testing needed before it can be merged/enabled.